backend: moved abs/avg plot data processing to func
